src/app/requirements.txt
psycopg==3.1.18
Django==5.0.3
djangorestframework==3.15.1
django-extensions==3.2.3
pytils==0.4.1
pillow==10.3.0
django-view-breadcrumbs==2.5.1
django-redis==3.1.6
django-debug-toolbar==4.3.0
django-ckeditor-5==0.2.12
django-browser-reload==1.12.1
django-watchman==1.3.0
django-uuslug==2.0.0
django-mptt-admin==2.6.2
uwsgi==2.0.25.1